Detection pyenv, rbenv, and plenv more robustly
We don't rely on the existence of those commands anymore. Instead it is enough
for the .rbenv, .pyenv, and .plenv to exists. Context: I used to rely on
homebrew to install those but this is a big dependence for something as crucial
as those scripts. So from now on, I shall install them directly by cloning
their repositories on github.